delete user

+++ src/main/java/com/planningpoker/service/UserService.java
@@ -38,7 +38,13 @@ public class UserService {
         return newUser;
     }
 
-    public void deleteUser(ObjectId id) {
-        userRepository.deleteById(id);
+    public boolean deleteUser(ObjectId id) {
+        if (userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+            System.out.println("Deleted user " + id + "!");
+            return true;
+        } else {
+            return false;
+        }
     }
 }
